Do not divide INGRID by area in DO_REGRID_A2A
Bug fix: in routine DO_REGRID_A2A, the INGRID (data
on the input grid) argument had been declared with
INTENT(INOUT).  Within the subroutine, INGRID could be
divided by the surface area.  This modifies the values
of INGRID that get sent back to the calling routine.

In some routines (i.e. the EDGAR routines), the data
array that was being passed to DO_REGRID_A2A was having
its values modified.  This could lead to anomalous
results upstream in the code.

The solution is to declare INGRID as INTENT(IN) and
then copy its values to a local shadow variable within
the subroutine.  Thus only the local variable is modified.

Also, we had forgotten to correct the indices in the
call to DO_REGRID_A2A for the EDGAR ship NOx.  This has
now been updated.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>