Add kludge in READ_AIRCRAFT_SO2 for NA nested grids
KLUDGE FOR NORTH AMERICAN NESTED GRID (bmy, 7/2/12)
For some reason there seems to be bad data within the N. American
nested grid data files.  Some of the longitude indices are out of
bounds for the grid.  Put in a simple kludge to just skip over
these boxes.  We are going to be reinventing how emissions get
done in GEOS-Chem, so it's probably not worth recreating the file
at this time.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>