Read new TOMS O3 data only for GEOS-5.7.2 met
The SEAC4RS code has a need to use the new data in the
TOMS_201203 directory.  However, as of 7/2/12, this update has
not yet been benchmarked in the std code.

For the time being, only read data from TOMS_201203 when using
GEOS-5.7.2 met.  Remove this criterion after we finally get a
chance to benchmark the TOMS_201203 data in the std code
(probably in the version after v9-01-03).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>