Added kludge to shut off bromine file I/O, esp. for nested grids
The current implementation of the bromine code assumes that
certain files will be read regardless of the settings of the
switches in input.geos.

schem.F: Do not read the Bry_Stratosphere_day.bpch and
Bry_Stratosphere_night.bpch files if Br_SCALING=0.

bromocarb_mod.F: Turn off reading the Warwick VSL data
in routine INIT_BROMOCARB if either LWARWICK_VSLS=.FALSE.
or Br_SCALING=0.

It is necessary to add this kludge for the nested grid, since
at present we only have Warwick VSL data for either 2 x 2.5
or 4 x 5.  A more consistent solution will be to read data from
the highest possible resolution and use MAP_A2A to regrid to
the current G-C resolution.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>