DO loop fixes in strat_chem_mod.F90
Rewrote several DO loops in strat_chem_mod.F90 such as

   DO I = 1, IGLOB
   DO J = 1, JGLOB
      ... etc ...

to be in the proper order, that is:

   DO J = 1, JGLOB
   DO I = 1, IGLOB
     ... etc ...

This allows the code to step thru the array in the same order
as it is laid out in memory.  This is more efficient and should
generate less cache misses.

Also, for the GEOS-5/GCM, we can no longer assume a minimum
tropopause level.  Always loop over the entire column
(L=1,LLPAR).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>