GET_NAME_EXT_2D returns "geos" for MERRA meteorology
GET_NAME_EXT_2D was written to returns the 1st 4 characters of
the grid name (i.e. "geos" instead of "geos4" or "geos5").  This
denotes data that is only defined at the surface.

However, for MERRA meteorology, GET_NAME_EXT_2D will truncate
the name "merra" to "merr".  Add as special case to replace this
with "geos".

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>