Move calls to findFreeLUN out of DEVEL blocks
Routine findFreeLUN in GeosUtil/inquire_mod.F90 is used to locate
file LUNs that are unopened.  This is because the GEOS-5 GCM
cannot use pre-defined (PARAMETER) file LUNs.

Move calls to findFreeLUN out of the DEVEL block and into the
main GEOS-Chem code.  For now, we have only done this in routines
that are called for the chemistry solver (and related operations).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>