Replace IU_* variables in file_mod.F w/ local variables
The IU_* parameters in GeosUtil/file_mod.F are incompatible with the
GEOS-5 GCM.  Therefore, we have rewritten several routines so that
these file units are stored in local varables or private module
variables.

The only variable that still needs to be left in file_mod.F is
IU_BPCH.  This is because IU_BPCH is used by several different
routines.  However, we have now made IU_BPCH a variable and not
a parameter.  IU_BPCH is initialized in READ_INPUT_FILE (in
module input_mod.F).

We have also deleted the IU_* parameters (except for IU_BPCH)
from file_mod.F.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>