Change the range of LUNs from 11..199 in inquireMod.F90
In routine findFreeLUN (in GeosUtil/inquireMod.F90), change the
range of valid LUNs from 11 to 199.  We had to start at 105 before
while so as not to conflict with the IU_* parameters in file_mod.F.
However, now that these IU_* parameters have been removed, we can
start the LUNs at the lower # of 11.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>