Bug fix for DEVEL: Now initialize PJ properly when using main.F
When compiling the Grid-Independent GC, only use the GCM
pressures to define the PJ array in FAST-J if we have selected
CPP switches EXTERNAL_GRID or EXTERNAL_FORCING.

Do not use the GCM pressures if DEVEL is selected.  This will
prevent allocation and/or out-of-bounds errors if you are compiling
with the "traditional" driver main.F instead of with the ESMF
interface.

Also add EXTERNAL_GRID and EXTERNAL_FORCING to a couple of unrelated
#ifdefs elsewhere in the code.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>