Bug fix in DEVEL: Only use GCM T w/ EXTERNAL_GRID, EXTERNAL_FORCING
Minor fix in GeosCore/set_prof.F.  Only rewrite the TJ array with
the GCM temperature (via the "T" array) if we compile G-C with the
EXTERNAL_GRID or EXTERNAL_FORCING Cpp switches.

As with the previous commit, you can compile G-C with the DEVEL
Cpp switch even if you do not use the ESMF interface.  Making this
update will help to ensure that we get the same results when
comparing the DEVEL simulation to a non-DEVEL simulation (for
testing/debugging purposes).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>