Remove potential parallelization bottleneck in calcrate.F
When the Bromine chemistry updates were added, a DO loop over all
lons/lats/levels (to initialize the QICE and QLIQ arrays) was placed
within subroutine calcrate.F.  But calcrate.F itself is called from
within a parallel loop from a higher level routine.  This means that
calcrate.F only works on a few grid boxes at a time (i.e. KBLOOP = 24
grid boxes), and not the whole globe.  This extra DO loop is therefore
unnecessary.

The extra iterations DO NOT appear to impact the scientific results,
but they do add extra CPU cycles to the simulation, which negatively
impacts the scalability.

We have rewritten the QICE and QLIQ arrays so that they are
dimensioned with KBLOOP instead of IIPAR,JJPAR,LLPAR.

We have also disabled a debug print statement in the routine
cldice_HbrHOBr_rxn.F.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>