Fixed a compilation error for Intel Fortran Compiler 12
Prasad Kasibhatla encountered a compilation error when using the
Intel Fortran Compiler (aka IFORT) version 12.  The error was:

   ifort -cpp -w -O2 -auto -noalign -convert big_endian -vec-report0
   -mcmodel=medium -i-dynamic -fp-model source -openmp -Dmultitask
   -I../Headers -module ../mod -I/opt/geos-netcdf-4/include -c -free
   strat_chem_mod.F90
   : catastrophic error: **Internal compiler error: segmentation violation signal raised** Please report this error along with the circumstances in which it occurred in a Software Problem Report.  Note: File and line given may not be explicit cause of this error.
   compilation aborted for strat_chem_mod.F90 (code 1)

Bob Yantosca determined that this error was caused by an OpenMP DO loop
in routine CALC_STE of strat_chem_mod.F90.  This error only occurs when
compiling with IFORT 12 for a nested-grid simulation at 0.5 x 0.666
resolution.  The error did not happen when compiling for 2 x 2.5 or
4 x 5 simulations.

It so happens that the the offending OpenMP DO loop does not execute for
nested-grid simulations.  Therefore, the quick fix is to extend the #if
statement to exclude this code from being compiled into the nested
grid executables.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>