Bug fixes to get DEVEL code to compile
Throughout DEVEL code, needed to add the argument am_I_Root to routines
that were updated in v9-01-03 for the grid-independent GEOS-Chem code.

Arrays in vdiffdr_mod.F90 were made pointers in v9-01-03 to facilitate
flipping arrays in the vertical. Fixed DEVEL code so that it is
consistent with this update.

Also added some missing dependencies in the Makefiles.

Signed-off-by: Melissa Payer <mpayer@seas.harvard.edu>