Prevent array out-of-bounds in tpcore_bc_mod.F
Sajeev Philip reported an array out-of-bounds error in array BC
(tpcore_bc_mod.F) when running the nested grid simulation. This error only
occurred when running the China nested grid simulation with the option to
save out TPCORE BC's over China (LWINDO_CH) set to FALSE in input.geos.

To avoid this error, we can remove the IF THEN statements preceding
the code to define the extent of the coarse-grid CH and SE BC regions so
that IM_BC_* are always defined.

Signed-off-by: Melissa Payer <mpayer@seas.harvard.edu>