Prevent out-of-bounds errors for ND46 diag w/ tagged CO simulation
In emissdr.F, the ND46 diagnostic was modified to include CH3Br
and CH2Br2 emissions.  However, the way this was implemented caused
an out-of-bounds error if IDECH3Br and IDECH2Br2 are zero.

The quick fix is to wrap these statements in IF statements to make sure
that IDECH3Br > 0 and IDECH2Br2 > 0.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>