Bug fix: pass am_I_Root=.TRUE. to GET_GLOBAL_CH4 in tagged_co_mod.F
GET_GLOBAL_CH4 was modified to pass the am_I_Root argument, which
restricts log file output to the root CPU.  This was a modification
for the grid-independent GEOS-Chem code.

We had neglected to pass this argument for the tagged CO simulation.
Therefore modify the call to GET_GLOBAL_CH4 from tagged_co_mod.F
by adding a .TRUE. value as the last argument.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>