Now call routine SET_COMMSOIL_MOD for all simulations
If we compile with -DDEVEL, -DEXTERNAL_GRID or -DEXTERNAL_FORCING, then
we call SET_COMMSOIL_MOD from routine GIGC_ALLOCATE_ALL (in module file
GeosCore/gigc_environment_mod.F90).  All arrays that depend on lon & lat
will be allocated with the local values of IIPAR, JJPAR, and then set
to zero.

If use the traditional GEOS-Chem, then call SET_COMMSOIL_MOD from
main.F.  In this case the arrays are not allocatable but defined with
the IIPAR,JJPAR parameters.  We set them to zero but do not allocate
them.

Eventually this code will be combined so that we keep the allocatable
arrays and remove the "traditional GEOS-Chem" code blocks.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>