Rename all LOCAL_MET to State_Met
In DEVEL code, renamed all occurrences of LOCAL_MET to State_Met so
that it is consistent with updates made in v9-02b. Also removed
references to gc_type_mod.

Additional met fields were added to Headers/gigc_state_mod.F90 so
that most of the met fields in dao_mod.F are now included in the
State_Met derived type object.

Signed-off-by: Melissa Payer <mpayer@seas.harvard.edu>