Replace dao_mod.F arrays w/ State_Met; start standardizing Input_Opt
We have now begun to replace the arrays in dao_mod.F with the pointer
fields of the State_Met object.  We need to pick State_Met for the ESMF
interface, and we cannot have the arrays in dao_mod.F declared
simultaneously due to memory usage.

Also, start bringing the Input Options object (Input_Opt) out of
the #if defined( DEVEL ) tags.  We will start using this in the standard
G-C code.  There will be many places in the code where we have to make
this modification.  We will bring Input_Opt out of DEVEL in a piecemeal
fashion so that we can test the implementation.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>