Now pass objects to INIT_STRAT_CHEM (strat_chem_mod.F90)
The call to routine INIT_STRAT_CHEM (in GeosCore/strat_chem_mod.F90)
has now been changed from:

   CALL INIT_STRAT_CHEM( am_I_Root )

to

   CALL INIT_STRAT_CHEM( am_I_Root, Input_Opt, State_Chm, RC )

so that we can now  pass the Input Options object and Chemistry State
object to the routine.  We also get back an error status flag RC that
will eventually be used for error handling.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>