Apply State_Met and Input_Opt updates to POPs code
The arrays in dao_mod.F have been replaced with the pointer fields of
the State_Met object. The POPs code has been updated accordingly.

Also updated the POPs code in input_mod.F so that options from the POPs
menu of input.geos are passed to the Input_Opt object.

Signed-off-by: Melissa Payer <mpayer@seas.harvard.edu>