Now undefine all grids when EXTERNAL_GRID, EXTERNAL_FORCING are used
Improved the error check in Headers/define.h.  If EXTERNAL_GRID or
EXTERNAL_FORCING are set, then we undefine all of the C-preprocessor
switches that specify met field grids (i.e. GEOS_4, GEOS_5, MERRA, etc.).

This is necessary to prevent wrongly defining the grid when running
GEOS-Chem in the ESMF environment.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>