Add gigc_input_opt_mod.o to dependency for comode_loop_mod.F
The routine INIT_COMODE_LOOP (in Headers/comode_loop_mod.F) now takes
the Input_Opt object as an argument.  The derived type (OptInput) that
is used to describe this object is contained in module file
Headers/gigc_input_opt_mod.F90.

Therefore, we need to add gigc_input_opt_mod.o as one of the object
files that comode_loop_mod.F depends on in the Dependencies section
of Headers/Makefile.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>