Replace STT with State_Chm
To facilitate the ESMF interface to GEOS-Chem, we have replaced the STT
array with State_Chm%Tracers. STT was made into a local pointer that points
to the State_Chm object field.

   STT => State_Chm%Tracers

During development, this update will remain in DEVEL tags. Once the updates
have been tested to ensure no changes to the science have been introduced,
we will remove the old code.

Signed-off-by: Melissa Payer <mpayer@seas.harvard.edu>