When using ESMF, do not call IS_LAST_DAY_GOOD from input_mod.F
When we connect to the GEOS-5 GCM via ESMF, we GEOS-5's History
Component handle all timing and diagnostic output.  GEOS-Chem diagnostic
output is not called.  Therefore we do not need to make sure that we
have scheduled output for the last day of the run.  This will help to
avoid premature exits when running the GEOS-5 GCM.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>