Additional bug fixes for GEOS-5 GCM interface via ESMF
ESMF/gigc_initialization_mod.F90: Needed to move the defintion of the
variable prtDebug to the top of the routine.  It was being used before
it was defined.

GeosCore/input_mod.F: In routine READ_CONVECTION_MENU, set TS_CONV to
TS_DYN when connecting to the GEOS-5 GCM.  This will prevent an error
trap in CHECK_TIMESTEPS from stopping the run.

GeosCore/physproc.F: Set RC = GIGC_SUCCESS.  Previously the RC argument
was not given a value.  This can trigger a compiler warning.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>