Bug fix: Make sure to use fields of Input_Opt whenever possible
Because we are broadcasting the fields of Input_Opt to other CPUs, we
should use those fields wherever possible.  For example, the function
ITS_A_FULLCHEM_SIM() from tracer_mod.F seemed to not be defined across
all CPUs, but the field Input_Opt%ITS_A_FULLCHEM_SIM was.

Also restore the drydep initialization for non-root CPUs in
ESMF/gigc_initialization_mod.F90.  We had moved this elsewhere but it
was incorrect.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>