Now use fields from Input_Opt instead of from logical_mod.F, etc.
We need to be careful to use fields from Input_Opt instead of the
corresponding fields from logical_mod.F, tracer_mod.F, etc.   When
connecting to the GEOS_5 GCM via ESMF, we do an MPI broadcast of all
of the fields of the Input_Opt object.  However, we do not broadcast
the corresponding fields in the older modules.  This can cause entire
sections of code not to execute.

Therefore, we now ensure that we use

  Input_Opt%ITS_A_FULLCHEM_SIM instead of ITS_A_FULLCHEM_SIM()
  Input_Opt%ITS_AN_AEROSOL_SIM instead of ITS_AN_AEROSOL_SIM()
  Input_Opt%LCHEM              instead of LCHEM  from logical_mod.F
  Input_Opt%LNLPBL             instead of LNLPBL from logical_mod.F

etc.  This has meant passing down Input_Opt to various subroutines where
it had not yet been passed before.

We also now call INIT_AEROSOL from routine GIGC_Init_Extra in
GeosCore/input_mod.F.  This moves the call out of the run phase and into
the init phase.

Also made several cosmetic changes in various routines.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>