Add Input_Opt%myCpu field for passing the CPU # down to G-C routines
We have added an extra field (myCpu) to the Input Options (Input_Opt)
object in Headers/gigc_input_opt_mod.F90.  This will allow passing of
the CPU # from the GEOS-5 GCM down to GEOS-Chem routines.  This will
assist with debugging, as it will allow us to see on which CPU an
error is occurring.

NOTE: We do not need to MPI Broadcast the Input_Opt%myCpu field to
non-root CPUs.  We will set Input_Opt%myCpu directly from the Initialize,
Run, and Finalize routines in the GEOSCHEMchem_GridCompMod.F90 module.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>