Restore fix for vertical regridding of OH for offline simulations
A fix for the vertical regridding of OH for offline simulations was provided
by Chris Holmes and implemented in v9-01-02p. The fix was later removed in
v9-01-03j during updates for the grid-independent model. To make the code
ESMF compliant, global size parameter LGLOB was removed, which conflicted
with the code added for the aforementioned fix.

The updates to fix vertical regridding of OH for offline simulations have
now been restored into the traditional GEOS-Chem code. To ensure that these
updates do not conflict with the GEOS-5 GCM, we have wrapped the code in
CPP switches. For example:

   #if defined( ESMF_ ) || defined( EXTERNAL_GRID ) || defined( EXTERNAL_FORCING )
      ! Do nothing
   #else
      ... put the code here ...
   #endif

A bug fix for nested NA offline simulations was also added to GET_GLOBAL_OH
so that GEOS-Chem looks for the correct OH file. Without the fix, a global
file at 0.5x0.666 resolution was read, resulting in an array out-of-bounds
error.

Signed-off-by: Melissa Payer <mpayer@seas.harvard.edu>