Need to pass State_Chm to the temporary version of DO_STRAT_CHEM
In the temporary version of routine DO_STRAT_CHEM (in module file
GeosCore/strat_chem_mod.F90), we had forgotten to pass the State_Chm
object.  This makes the argument list consistent with the version of
DO_STRAT_CHEM that is called in the standard G-C (without ESMF).
Now fixed.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>