Fix for GEOS-5 GCM: avoid polar boxes with zero surface area
The GEOS-5 GCM is a grid-point model, with the polar boxes (+90 & -90
degrees latitude) being grid box centers.  But since GEOS-Chem also
needs to know the latitudes at the north & south edges of each grid box,
we have to make a kludge.

For all grid boxes along the NORTH POLE (+90 lat), we let the northern
edge be greater than 90 degrees.  For example, if the grid spacing is
1 degree in latitude, then for all longitudes at the North Pole:

  * The N. EDGE of each grid box is +90.5 degrees
  * The CENTER  of each grid box is +90   degrees
  * The S. EDGE of each grid box is +89.5 degrees.

Similarly, at for all grid boxes along the SOUTH POLE, we have this
condition (also assuming a grid spacing of 1 degree in latitude):

  * The N. EDGE of each grid box is -89.5 degrees
  * The CENTER  of each grid box is -90   degrees
  * The S. EDGE of each grid box is -90.5 degrees.

Therefore, at the poles, the latitudes at the northern and southern
edges of each grid box are symmetric around either +90 degrees or -90
degrees.  When you take the difference of the sine of the latitudes
at the north and south edges of a polar grid box, the terms will cancel
each other out, resulting in a grid box surface area that is zero.

We can take advantage of this symmetry around +90 and -90 degrees to
make a simple fix:

(1) AT THE SOUTH POLE: Subtract the sine of the latitude at the north
edge of the grid box from the sine of -90 degrees (which is -1) and
then multiply by 2.

(2) AT THE NORTH POLE: Subtract the sine of +90 degrees (which is 1)
from the sine of the latitude at the south edge of the grid box, and
then multiply by 2.

This fix avoids having polar grid boxes with zero area.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>