Remove DEVEL tags from State_Chm%TRACERS, State_Chm%SPECIES #1
We have started the process of mainling the replacement of the
STT array by State_Chm%TRACERS and the CSPEC_FULL array by
State_Chm%Species.  Begin the long process of removing the DEVEL
Cpp blocks from around these sections.

We have also consolidated argument lists with work that had been done
previously on the ESMF branch.  We also now are starting to replace
variables from logical_mod.F and tracer_mod.F with the corresponding
values in the Input_Opt object.  This is necessary for the GEOS-5 GCM
interface, since the values in Input_Opt need to be read on one CPU
and then broadcasted to the remaining CPUs.

Also moved some obsolete code to the obsolete/ directory.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>