Ship NOx emissions: Add NO2 to Ox emissions
If we are using the non-local PBL mixing, then emissions
occur when O3 and NO2 concentrations are lumped into the
Ox family, so we add the ship NO2 emissions to the O3
emissions to get Ox emissions.

When using instant PBL mixing, then emissions occur in the
chemical solver, where there is no chemical family lumping.
Then O3 and NO2 production should not be combined.

Signed-off-by: Chris Holmes <cdholmes@uci.edu>