Hg sim: STT contains RGM and PBM
The previous Hg2 and HgP tracers are now assumed to be
RGM and PBM, respectively. There is no refractory HgP.
All Hg(II) is assumed to partition between gas and aerosol.
Equilibrium partitioning is now calculated before and
after chemistry. Dry and wet deposition no longer need
special consideration of gas-aerosol partitioning.
Special diagnostics for RGM and PBM are no longer required
either, since they are contained in the ND45 diagnostics.

This update simplifies the source code, but does not affect
how the model operates. Surface Hg(0) changes <5%. RGM and
PBM, change up to 25%.
Currently, PBM is not susceptible to
reduction, until it later enters gas phase but this could
be changed.

Bugs in gamap_mod that affected the tracerinfo.dat are
fixed.

Signed-off-by: Chris Holmes <cdholmes@uci.edu>