Add KPP tolerance fix when DEVEL=yes (mostly for GEOS-5 GCM)
Mike Long has added a tolerance fix when we are using the KPP solver.
This fix is only activated when you compile GEOS-Chem with DEVEL=yes.

When KPP encounters a box that it can't solve, instead of dying with
an error after 2 tries (which is the default behavior), we increase
the tolerances by a factor of 2 and try again.  This allows the solver
to move forward in the GEOS-GCM.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>