Add option KPP_SOLVE_ALWAYS=yes to skip past nonconvergence
In the previous commit, we added a fix that would cause KPP to loosen
tolerances if a non-convergent box were encountered.  This would let
a GEOS-Chem run continue instead of letting the run die after two
non-convergences.

We had originally activated this fix when GEOS-Chem is compiled with
the DEVEL=yes option.  We have now extended this so that the fix will
also be activated if G-C is compiled with KPP_SOLVE_ALWAYS=yes.  This
will let the fix be activated without turning on additional DEVEL
code.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>