ND63 bugfix: Write diagnostics when run doesn't start at 00:00
Previously the ND63 diagnostic file was opened at the start
of each new day. When runs started at a time other than 00:00
the diagnostic file was never opened and a I/O error occured.
Now we open the diagnostic file at the start of the run,
regardless of the time of day.

Similar errors may occur with other timeseries diagnostics
(ND48, ND49, ND50, ND51), but I have not checked.

Signed-off-by: Chris Holmes <cdholmes@uci.edu>