Debug tracking
Problems with bromine and heterogeneous chemistry persist, but in
the interests of tracking changes to the code, making an interim
commit before starting work on APM upgrades. Aerosol formation
code has now been doubly and triply verified; also added code
to ensure that stratospheric bromine is set to reasonable values
(if an input switch is set).

Signed-off-by: Sebastian David Eastham <seastham@mit.edu>