ND32, ND36 bugfix:
Ship emissions of NOx, O3 and HNO3 are included in the
ND32 and ND36 diagnostics. However, ship plume chemistry
and diagnostics are calculated every convective time step
when using non-linear PBL mixing, rather than at the emission
time step, which caused errors in the diagnostics. We now
add counters for the ship emission diagnostics so that they
can be correctly aggregated with other emission diagnostics.

Signed-off-by: Chris Holmes <cdholmes@uci.edu>