Corrected error in SLA calculation
When calculating SLA properties, take liquid H2SO4 as total
available, but was bounds-checking total rather than liquid.
As a result was not preventing errors in regions where the
total H2SO4 was sufficient for SLA formation but not the liquid
H2SO4. This was corrected. Also improved stability-checking of
heterogeneous chemistry.

Signed-off-by: Sebastian David Eastham <seastham@mit.edu>