Now pass State_Chm and Input_Opt to TOMAS routines (and calling routines)
The last state of the TOMAS code used STT from tracer_mod.F and
the various logicals from input_mod.F.  But we have now removed all
references to these variables as part of the grid-independent updates.
Therefore, we now pass State_Chm and Input_Opt to all TOMAS routines
where they are needed.

In various locations in the code, we also now point to the TRACERS field
of the State_Chm object:

   STT => State_Chm%TRACERS
   ...
   NULLIFY(STT)

in order to preserve historical code without making too many modifications.

We get identical results when compiling with DEVEL=yes and DEVEL=no.
We also verified that the code compiles for TOMAS=yes and TOMAS40=yes
options, but we invite the TOMAS team to do further testing.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>