Add -check arg_temp_created option to DEBUG=yes Makefile switch
When we compile GEOS-Chem with array checking, we will now also check
to see if array temporaries are created.  These are internal copies of
arrays that get created when passing data to subroutines.  Array
temporaries can adversely impact GEOS-Chem's performance, so we will
take steps to detect them when compiling with DEBUG=yes.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>