Add SALT_MASS to !OMP PRIVATE declaration in seasalt_mod.F
In routine WET_SETTLING (in GeosCore/seasalt_mod.F), variable SALT_MASS
should have been added to the !$OMP PRIVATE list but wasn't.  This caused
numerical differences that manifested themselves when DEBUG=yes was
invoked.

We have now added SALT_MASS to the !$OMP PRIVATE list, and this
corrects the problem.   We have also reformatted some of the comments
in routine WET_SETTLING, for clarity.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>