Use wind PDF to calculate dust emission and read new source map
Two changes are implemented in the file:
1) the dust emission is integrated over a weibull distribution
constructed from the .w1 met files containing the mean and
standard deviation of 10m wind. This removes some of the
resolution dependence of the dust emissions. NB In this version
the scaling factor for dust emissions has been fixed for all
resolutions.
2) a new source map (generated and stored in the same file as the
original GOCART source map, but under tracer 88) is used based on
surface roughness and levelness (Koven et al. 2008). This allows
for emissions from vegetated regions (the original GOCART source map
was very conservative). Emissions are modulated by the bareness fraction
derived from NDVI on a monthly basis from 1983-2006 (Kim et al., 2013).
For simulations outside of this time range either 1983 or 2006 vegetation
is used.

Signed-off-by: David A Ridley <daridley@mit.edu>