New jv_spec_aod.dat files read and include calc_aod
The new jv_spec_aod.dat is split into separate files for
each species, currently still in text format but generated
from a master NETCDF file that will be read directly eventually.
The new files contain properties at 61 wavelengths, the original
wavelengths and also those for RRTMG. These are interpolated from
GADS and updated where improvements are reported in literature.

Subroutine CALC_AOD works out the closest tie points
in the optics LUT wavelengths and the coefficients required to
calculate the angstrom exponent for interpolating to the requested
wavelength. If the wavelength requested matches a standard wavelength
in the LUT then we skip the interpolation

Signed-off-by: David A Ridley <daridley@mit.edu>