Add SOA tracers to aerosol mass for fastj
Add simplified SOA tracers to total OCPI aerosol mass to be used in fastj.
This is how it is done with the more complicated aerosol species as well.
Note that for now we have assumed the same (SOA mass:carbon mass) ratio as
for OC, but this is probably underestimated and should likely be fixed in
future updates.

Signed-off-by: Jenny Fisher <jafisher@fas.harvard.edu>