Multiple wavelength input/output for optics
User can now specify any 1-3 wavelengths in the input.geos.rad
file for output. These will be output in the OD-MAP-$ diagnostic and
also in the RRTMG RADMAP-$ diagnostic if RRTMG is switched on.
For ND48-51b diagnostics the first wavelength specified is the one
that is output.

Possible O3 output error in previous version of this branch
has been corrected (fields may have been coming out as zero.

Signed-off-by: David A Ridley <daridley@mit.edu>