User-selected wavelength for all diagnostic outputs and ODAER resize
***NOTE - THIS CODE MAY SCREW UP PHOTOLYSIS (SEE BELOW)

Diagnostics 48-51b now output AOD at the first wavelength selected
by the user in input.geos.rad, even if it requires interpolation.
-this should work

NAER*NRH fields are stored in ODAER to retain the IRH for OPMIE.F
but with the added wavelengths this is a huge 5D array...
The code was altered just to store the relevant AOD and not the
IRH so that the size could be reduced to NAER arrays.
However, the photolysis code OPMIE reads the old jv_spec.dat and expects
the ODAER to store the OD in the right RH bins - so at the moment
all data is stored in the first 5 bins and the photolysis code will
therefore read the properties for SO4 at each wavelength (instead of
the properties for the different aerosol at the relevant wavelength).

A possible solution is to read aerosol from the new .dat files...
jv_spec.dat aerosol is redundant now we have wavelength bins
in the new .dat files. Need to rewrite OPMIE.F to use those values. \
We don't want to have to calculate
the RH interpolation twice so need to think of a way to map that
between aerosol_mod and opmie...

For now this has the issue but *shouldn't* affect results too dramatically.

Signed-off-by: David A Ridley <daridley@mit.edu>