Elimate duplicate IF statement for LRETRO in biofuel_mod.F
The Git merge of v9-02l with a branch containing the RCP emissions updates
(based on v9-01-02n) resulted in duplicate IF statements for zeroing
biofuels when RETRO is turned on. We have removed the second IF statement
(from the RCP version of the code) because it still included zeroing
biofuels for C2H6. We have since removed replaced RETRO C2H6 emissions with
Yaping Xiao's anthropogenic C2H6 emissions, which do not include biofuel.
Therefore, we should not zero biofuels for C2H6 here.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>