Correct OOB error in updated volcano emissions code
In the volcano emissions routines, arrays vnv_ii (index of longitudes
of non-erupting volcanoes) was found to have values greater than IIPAR
when running the nested-grid simulation at 0.25 x 0.3125.

The recommended fix is to place error traps that will cause the code
to skip boxes where vnv_ii and vnv_jj have values that are not in the
range 1..IIPAR and 1..JJPAR.

Ditto for the arrays vev_ii and vev_jj (which are the lon & lat of
eruptive volcanoes).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>