Photolysis AOD fix and sea-salt SO4-NIT addition
Photolysis code wasnt receiving AOD columns correctly.
Now gets those at 1000nm and put in the correct RH bins so the
right optics can be read out of jv_spec.dat.
SO4s and NITs tracers now included in SNA for optics.
Still not associated with sea-salt (i.e. should be part of the
larger SS particles so optics will be different) but this is a
work in progress.

Signed-off-by: David A Ridley <daridley@mit.edu>