Removed "define.h" from everywhere in the code
The include file "define.h" is now obsolete.  We now rely on the
Makefile_header.mk to define the proper C-preprocessor switches for
GEOS-Chem, based on user inputs from the command line.

NOTE: This is a critical modification for the unit tester.  We need
to have a script be able to make the GEOS-Chem code repeatedly with
various debugging options within the unit tester framework.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>