LSTRATOD, LACSO2, mesospheric H2SO4 phot. and minor fixes
Added a switch controlling whether or not liquid strat.
aerosols are considered for photolysis, and the non-AEIC
aircraft SO2 emissions can now be enabled or disabled
independently of AEIC. H2SO4 photolysis is now extended
to the model top. Minor fixes which were implemented in
later versions of GEOS-Chem have been added here.

Signed-off-by: Sebastian David Eastham <seastham@mit.edu>