Fixed a numerical issue in GeosCore/paranox_mod.F
In routine INTERPOLATE_LUT2, there is a division JO1D / JNO2.  But
because JNO2 is a J-value, its value will go to zero at night.  This
will result in a div-by-zero error.

Geert Vinken recommends that if the division cannot be done, that we
just set the result of the quotient JO1D / JNO2 to zero.  We have
added a call to SAFE_DIV in GeosUtil/error_mod.F to do this.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>