Improved validation of command line inputs in Makefile_header.mk
Except for a few instances, we test all Makefile variables (whose values
are defined at the command line) with an ifeq statement.  This is
especially important for those variables having a yes/no answer.
Instead of relying on an ifdef (which will return TRUE if any value is
passed), we now look for a "y" or a "yes" value (case-insensitive).

This was done to facilitate automatic debugging of GEOS-Chem in the
unit tester.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>