Prevent floating-point invalid in drydep_mod.F
In routine DEPVEL, we shall skip to the next species when the molecular
weight for drydep species K (aka XMW(K)) is zero.  Some species are given
a zero molecular weight to denote that they shall use the deposition
velocity of another species.  For example, PMN uses the same dep. vel.
as PAN.

This error condition was caught by the GEOS-Chem unit tester when
compiling with FPE=yes.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>