Further fixes in drydep_mod.F
Add a better check for the floating-point invalid trap.  Now use the
LDEP logical array in routine DEPVEL to determine which species will
get drydep velocities computed and which won't.  Skip over species
that have both HSTAR=0 and XMW=0.  Both of these quantities are found
in denominators, which could lead to a div-by-zero condition.

Also, we now assign a XMW=118d-3 (118 g/mol) as the molecular weight
for RIP and IEPOX.  Formerly, XMW=0 for these species.  Fabien Paulot
said that this was a typo and that we should use the molecular weights
for RIP and IEPOX.  This also avoids div-by-zero conditions.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>