Now hold IW, ZO $OMP PRIVATE in GeosCore/drydep_mod.F
We have explicitly added IW to the !$OMP+PRIVATE list in routine DEPVEL
in GeosCore/drydep_mod.F.  This should already be held private because it
is a loop index, but it doesn't hurt to add it to the list.

We also have added ZO as a PRIVATE variable when compiling with GCAP
meteorology.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>