Prevent out-of-bounds error in AIRQNT_FULLGRID for Hg simulations
In routine AIRQNT_FULLGRID (in GeosCore/dao_mod.F), we loop over 1..LGLOB
vertical levels.  However, when we compile with -DGRIDREDUCED, the
AREA_M2 array in GeosUtil/grid_mod.F90 will only be compiled with
LLPAR vertical levels.  This means that the call GET_AREA_M2( I, J, L )
will be trying to retrieve data from AREA_M2 with a vertical level index
that is out-of-bounds.

The quick solution is to just call GET_AREA_M2( I, J, 1 ), since for
the traditional GEOS-Chem, all areas in a vertical column are identical.
In fact, the (I,J,L) dependence on AREA_M2 was introduced as part of the
Grid-Independent GEOS-Chem in order to allow for future grids that may
not be rectilinear.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>