In Makefile_header.mk, now evaluate GIGC inputs first
In the Makefile_header.mk file, we now evaluate the command-line inputs
for the Grid-Independent GEOS-Chem first.  If the user compiles GEOS-Chem
with one of the following combinations:

  make -j4 ... ESMF=yes ...
  make -j4 ... ESMF=yes EXTERNAL_FORCING=yes ...
  make -j4 ... ESMF=yes EXTERNAL_GRID=yes ...

Then the make process will not require you to supply a value for GRID,
since in this instance we will be getting the grid supplied to us from
external context (e.g. from the GEOS-5 GCM).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>