Read lines of input.geos containing up to 500 chars
Because we have updated the number of tracers (especially for the SOA
simulations), we likewise had to increase the character string length
that is used to read text from the input.geos file.  In the DIAGNOSTIC
MENU section, it is possible to have a list of tracer numbers occupy
more than 300 characters.

Changed the MAXDIM parameter from 255 to 500.  We can change it further
in the future.  By changing MAXDIM, all other relevant character string
lengths are set accordingly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>